Rename DTO to value object.
